textfields: bring shape into view when editing
